+++ src/main/java/game/stage/StageSingle.java
@@ -12,7 +12,7 @@ public class StageSingle extends Stage {
 
     @Override
     public void init() {
-        friendlyArr[0] = new FriendlyGuided(new Point(WIDTH / 2 + Friendly.WIDTH / 2, HEIGHT -100));
+        friendlyArr[0] = new FriendlyGuided(new Point((double)WIDTH / 2 + (double)Friendly.WIDTH / 2, HEIGHT -100));
         friendlyArr[0].setBulletLevel(1);
 //        friendlyArr[1] = new SatelliteLeft(friendlyArr[0]);
 //        friendlyArr[2] = new SatelliteRight(friendlyArr[0]);
